Chore: moves localstorage helper as a shared utility and refactors constants
