+++ app/javascript/dashboard/api/testimonials.js
@@ -1,5 +1,5 @@
 /* global axios */
-import wootConstants from 'dashboard/constants';
+import wootConstants from 'dashboard/constants/globals';
 
 export const getTestimonialContent = () => {
   return axios.get(wootConstants.TESTIMONIAL_URL);
