--- app/javascript/dashboard/constants.js
@@ -1,28 +1,0 @@
-export default {
-  GRAVATAR_URL: 'https://www.gravatar.com/avatar/',
-  ASSIGNEE_TYPE: {
-    ME: 'me',
-    UNASSIGNED: 'unassigned',
-    ALL: 'all',
-  },
-  STATUS_TYPE: {
-    OPEN: 'open',
-    RESOLVED: 'resolved',
-    PENDING: 'pending',
-    SNOOZED: 'snoozed',
-    ALL: 'all',
-  },
-  ARTICLE_STATUS_TYPES: {
-    DRAFT: 0,
-    PUBLISH: 1,
-    ARCHIVE: 2,
-  },
-  LAYOUT_TYPES: {
-    CONDENSED: 'condensed',
-    EXPANDED: 'expanded',
-  },
-  DOCS_URL: '//www.chatwoot.com/docs/product/',
-  TESTIMONIAL_URL: 'https://testimonials.cdn.chatwoot.com/content.json',
-  SMALL_SCREEN_BREAKPOINT: 1024,
-};
-export const DEFAULT_REDIRECT_URL = '/app/';
