+++ app/javascript/dashboard/constants/localStorage.js
@@ -1,0 +1,5 @@
+export const LOCAL_STORAGE_KEYS = {
+  DISMISSED_UPDATES: 'dismissedUpdates',
+  WIDGET_BUILDER: 'widgetBubble_',
+  DRAFT_MESSAGES: 'draftMessages',
+};
