+++ app/javascript/shared/helpers/localStorage.js
@@ -1,9 +1,3 @@
-export const LOCAL_STORAGE_KEYS = {
-  DISMISSED_UPDATES: 'dismissedUpdates',
-  WIDGET_BUILDER: 'widgetBubble_',
-  DRAFT_MESSAGES: 'draftMessages',
-};
-
 export const LocalStorage = {
   clearAll() {
     window.localStorage.clear();
